move meta status and add alert manager
